Update offer viewmodels #118

+++ Data.Abstractions/Models/IOffer.cs
@@ -8,7 +8,7 @@ namespace Data.Abstractions.Models
     {
         string Description { get; set; }
 
-        double Price { get; set; }
+        int Price { get; set; }
 
         int EstimatedTime { get; set; }
